Fix-ups to tests to reconcile scamper-instance renaming
